infinity: remove spinner when request completes but response is empty
Summary:
Reported in 1014-gh-p2

Current code clears the spinner when the infinity xhr call [[ https://[private link].js#363 | responds with an error ]], or when the request succeeds and the payload is processed completely.

When two xhr calls are made in quick succession, and one of the calls returns first with the final payload batch, the second call will also succeed but with a payload of `{"type":"empty"}`, which can result in the infinity spinner never being cleared.

 In this changeset, we cover the case described above.

Test Plan:
1. Test on a P2020 site with only a few posts -- i.e. not enough posts to fill the screen when in compact view.
2. Scroll your browser to the very bottom, and reload.
3. After the very last post (oldest) is loaded, the spinner should clear.
4. Test on Default/Expanded view, and also on non-P2 sites with infinity-scroll enabled themes, e.g. pub/twentytwenty. Infinity scroll should still behave well.

Reviewers: #lighthouse_team, undemian

Reviewed By: #lighthouse_team, undemian

Subscribers: undemian

Tags: #touches_jetpack_files

Differential Revision: D49344-code

This commit syncs r213450-wpcom.