Gutenboarding: save selected features as a site option
Summary:
This patch is part of [[ 45148-gh-wp-calypso | #45148 ]]

This patch saves the `selected_features` as a site option.
We're looking to use this data during the Site setup flow (in editor) to recommend a plan based on the selected features.
The option will be available only for sites created using `site_creation_flow: 'gutenboarding'`.

Note: this change is a very similar to D42191-code

Test Plan: Follow testing instructions in [[ 45148-gh-wp-calypso | #45148 ]]

Reviewers: simison, mciampini, tomjcafferkey

Reviewed By: mciampini, tomjcafferkey

Tags: #touches_jetpack_files

Differential Revision: D49420-code

This commit syncs r213657-wpcom.