Module List: filter all items using existing filter
Fixes #17229

The existing jetpack_modules_list_table_items filter ended up
not being very useful until now. It was only filtering the items
displayed, not the whole list of items available on the page.
The filter consequently could not be used to stop a specific item
from being displayed.

With this change, one can now filter an item out entirely.
I also took the opportunity to apply the filter to the views as well,
so when you filter an item out, the total count of items is updated.