Calypsoify: Remove unnecessary redirect
Currently, if we've specified `?calypsoify=1` or `?calypsoify=0`, we redirect after we set the user meta.

While this works, a redirect here is really unnecessary, because we can just enable or disable the flag if this argument is set. 

A redirect was needed in the original internal plugin version, because it works with cookies; while here, we just set a user meta.