Contact Form: Allow for commas and such in field labels, options, and values
When converting the options and values arrays to shortcodes, it's just
joining the options as a comma-separated string. If any of the elements
of these arrays contain commas, this doesn't round-trip properly.

It also turns out that the shortcode parser is fairly limited, in
particular it does not handle attribute values containing brackets or
backslashes correctly (which precludes using JSON directly). The Contact
Form code isn't accounting for that either.

On the other hand, the shortcode parser works fine with HTML entities in
the attribute values, and we want this user input to be entity-encoded
anyway. So we can solve all of that in one go by entity-encoding all the
relevant characters.

This patch also adds some extra test improvements that I did while
trying to find an existing test to extend; I figured that since I did
the work, we may as well include it.

Fixes #7633