[not verified] Copy UnsortedUseStatementsSniff from MediaWiki
After a merge of two patches adding `use` statements for the same class
in different places in the existing list broke things, discussion on
Slack was in favor of adding a sniff to require that `use` statements
are in alphabetical order.

As https://github.com/wikimedia/mediawiki-tools-codesniffer already has
just such a sniff, we'll copy theirs and modify it to suit our purposes.

This commit copies it with minimal changes for compatibility with PHP 5.6
(as detected by `composer php:compatibility`), with the changes to make
it pass all our coding conventions to be done as a followup because I
want to see just how many changes that requires.