Codesniffer: Add a package to hold our coding standard
This will make it a bit cleaner in the future to manage tests and
dependencies for the coding standard.

And, if some other repo wants to use our coding standards, they can more
easily do so.

The new coding standard package imports the PHPCompatibility and
WordPress standards, so those are removed from `.phpcs.xml.dist`.
The phpcs configuration specific to Jetpack remains in
`.phpcs.xml.dist`.