Editor Metadata: Add the last editor used to the post meta
It some circumstances it is beneficial to know which editor has been
used to edit a post. The classic editor plugin already stores this
information so that it can be displayed in the post list, and be used to
initialise the correct editor by default.

This adds the same functionality to Jetpack, and adds the metadata to
the post if the classic editor plugin will not, either because it's not
active, or because of it's settings.