Fix memory and performance issues of the Widget Visibility
The Widget Visibility feature requires a lot of data in order to provide a very flexible visibility settings for the widgets. For instance, it lists all the pages on a site in a dropdown, which may be an issue for sites with a huge amount of pages, or sites with very large pages, so they don't fit into a single memcache key=>value pair.

In order to preserve the current behaviour (listing all the pages), but optimize the loading times, this patch is introducing a custom direct SQL query for fetching only necessary data of all pages (ID, post_parent, and post_title). The direct SQL query is being cached under plugin's own group, and is taking advantage of the `wp_cache_get_last_changed( 'posts' )` value, which is being flushed whenever a new page is being added, so the listing should be always fresh (the `wp_cache_get_last_changed( 'posts' )` is being used also inside the `get_pages` function call, which the custom implementation is replacing).

The custom `get_pages` function does not trigger the memcache gets for all the pages, as we really don't need whole post objects here.

This changeset also limits the fields returned by the `get_users` call to `ID` and `display_name`, since that's really all we need here.

Fixes #6955