Stats: Insist on a WP_Post when trying to fetch a post ID
It's completely unclear what might be happening in #15582 or how it
might be reproduced. But some static analysis (see the issue) seems to
indicate that Stats wants a WP_Post object and not anything else, so
perhaps we can cut the spaghetti knot and just check for that class
directly in Stats.

Also, since phpcs is making me touch the surrounding lines anyway for
alignment of `=`, remove some extraneous parens and throw in some
try-finally logic to make it a little more robust.

Fixes #15582