Add the `get_pages` filter
In order to offer full backward compatibility with original `get_pages()` function call, the `get_pages` filter is being added to the custom `get_pages` method. In order to be able to provide the 2nd argument ($parsed_args), I had to copy-paste the array of default args from the `get_pages()` function.