[not verified] Test Embed API: Update resolve-redirect route
While testing on a local dev site I noticed that the resolve-redirect
API call was being 301 redirected and would then fail as the `//` in the
URL parameter was being switched to a single `/`.

After a lot of debugging I discovered that it was Cloudflare that was
seeing a URL like
`https://test.com/wp-json/wpcom/v2/resolve-redirect/https://calendly.com`
as invalid, and redirecting to a sanitised version.

The solution was to make the URL parameter in the path optional, so that
it could be passed as a query parameter.

This change implements that, so that the old style of calling the
endpoint will still work, but also we can swap to using a more
conventional URL format.