JITM: allow pre-connection JITMs to specify a list of required plugins
We're working on making the JITM package useful for plugins other than Jetpack. The current
pre-connection messages are all Jetpack specific. Let's add a 'plugin_active' key for the message
arrays that will be used to provide a list of required plugins. If at least one of the plugins in
the message's list is active, the JITM will be displayed.