Autoloader: Added a global to indicate when the autoloader is being included by another autoloader
There was an assumption made that the inclusion of an `autoload_packages.php` file indicates a plugin is "active". The problem here is that the cache mechanism means we might load an autoloader from a deactivated plugin, leading to it being seen as "active" and kept in the cache. This new global lets the autoloader detect when it is being included from an older one so that it can take appropriate steps.