Danger Bot: use code freeze date provided in milestone description
We currently automatically set the code freeze date displayed in Danger Bot messages, 7 days prior to the milestone's due date.

This holds true for most of the year, but in some scenarios we deviate from this one-week beta period standard.

When that's the case, we can define a code freeze date in the milestone description (using ISO 8601), and we'll now try to pick this up with Danger.
If no date has been provided there, we'll fallback to the previous behaviour.