Tests: Speed up linting with php-parallel-lint
We currently run `php -l` over most of the PHP files by having a PHPUnit
test that shells out to `find`. Which works, but it's pretty slow (about
35 seconds in my local environment).

We can improve it significantly using php-parallel-lint.[1] Locally for
me it's down to 2.8 seconds.

I've also moved the running of it out of PHPUnit and into a separate
script. I'd have added a composer alias for the new script too, but
"php:lint" was already taken for phpcs.

[1]: https://packagist.org/packages/php-parallel-lint/php-parallel-lint