CI: Add a Github workflow to run php linting
We seem to be moving away from Travis and to Github Actions. So, this
moves linting and PHPCompatibility out of Travis and into a Github
Action.

This also adds running of the full PHPCS Standard over the non-excluded
files, so we don't regress on them in the future as we have in the past.