Git hooks: Improve "not verified" flagging
Improvements:

* We can more reliably detect whether the hook was run by storing the
  tree hash that's going to be committed rather than a timestamp.
* If the commit is already flagged as not verified, there's no point in
  adding the flag again ("[not verified] [not verified]").
* If the commit is flagged as not verified but it was just verified, we
  can remove the flag.

There seems to be no way to do the right thing (i.e. verify) on a cherry
pick or rebase. But this does still improve the situation for those too.
For a cherry pick you can `git commit --amend` immediately after, which
will verify and will now remove the flag. For a rebase, you can do it
with --interactive and change "pick" to "reword" on all lines, which
will for some reason also verify each one.