Codesniffer: Fix code coverage generation hang due to Generic.PHP.Syntax sniff
We generate code coverage using phpdbg. If you pass it `-qrr` it works
to execute command-line scripts, but it's not really command-line
compatible with the standard PHP executable.

None the less, it still sets `PHP_BINARY` to point to itself, and
without the `-qrr`. This will break anything that tries to use
`PHP_BINARY` to shell out to PHP, and in particular breaks the
Generic.PHP.Syntax sniff's attempt to shell out to `php -l`.

So when we try to generate code coverage for the codesniffer package,
and it executes the sniffs to test them, the Generic.PHP.Syntax sniff
hangs.

There are a few ways we might work around this:

1. Exclude Generic.PHP.Syntax from the Jetpack standard. Linting can be
   done more efficiently with php-parallel-lint.
2. Exclude Generic.PHP.Syntax when running tests, as if phpcs had been
   invoked by `--exclude=Generic.PHP.Syntax`.
3. Have the tests override phpcs's use of `PHP_BINARY`. But,
   a. We'd have to determine what replacement to use.
   b. Phpcs does this in static methods with global state.

This patch does both #1 and #2. #3 isn't _hard_, but might be fragile so
let's avoid it unless we have to.