[not verified] Fix files checked for `php:compatibility`
We already have the PHPCS exclusion rules defined, let's split them out
so they can be used by the `composer php:compatibility` check too.

That revealed a few compatibility fails in some existing files that had
been falling through the cracks before.

Not verifying this commit to make a point, verification will happen in
the next commit in the PR.