Fix 1000 other phpcs sniffs
(that may be slightly hyperbole)

Touching a few lines to fix PHP Compatibility triggered phpcs-changed to
start complaining about existing crap throughout the rest of the files.
Sigh. I can't tell if it just complained up to the end of the file, or
if the errors were so thick that it cascaded.

If anyone wants to improve the low-effort "Test 'foo'" doc comments
being added here, feel free to do so. I spend enough time on this
already.