Git hooks: Revert code removing the "not verified" tag
In #17728 we added some code to remove the "not verified" tag when
verification passed, on the assumption that the pre-commit hook was
verifying the whole commit on an --amend.

It turns out that it only verifies the changes being made, rather than
the commit as a whole. So removing the tag is not the correct thing to
do unless we can figure out a way to detect --amend and reverify the
whole commit.