Update Jetpack PHPUnit tests for PHP 8
WordPress requires PHPUnit 7, which doesn't support PHP 8. To allow it
to run, WordPress core has a local version of the MockObject classes
that have been edited to work.

Then we also need to clean up various tests that fail in PHP 8. And I'll
throw in a few other cleanups that happened to show up while I was doing
that.

Note this PR doesn't fix package tests. They're different and involved
enough that I'll put them in a separate PR.