Fix WP_Test_Jetpack_Sync_Constants::test_does_not_fire_if_constants_havent_changed
It's calling `constant( $constant )` and trying to catch the PHPUnit
exception created from the PHP warning, but in PHP 8 it throws an
Error instead of generating a warning.

The whole idea here seems suspect, if any of the constants are not
defined the whole test gets skipped, which makes me wonder how it was
ever being run in the first place. More sensible would be to use
`defined()` to check each individual constant.

Also, may as well switch assertEquals() to assertSame() to avoid
potential false passes.