Build: use symlinked dev versions of packages in dev environment (#17835)
Composer 2.x sets all repositories as canonical by default. We added the
"canonical": false setting to our local path repo in composer.json so that
stable versions would be used when the release version of Jetpack is built.
However, this causes stable versions to also be used in the development
environment, which is undesirable. During development, the development
versions of the packages from the current branch should be symlinked to the
vendor directory.

To fix this, the "canonical": false setting should be removed from the
local path repo.

To prevent development versions from being used when the release version of
Jetpack is built, update version-packages.sh to remove the local repo from
composer.json.