Anti-spam: detect Akismet when it's installed as an mu plugin
The Akismet endpoints only detect the Akismet plugin status when Akismet is installed in the plugins
directory. Refactor the code to allow Akismet to be detected when it's installed in the plugins
directory or as an mu plugin.

Also, use the AKISMET__PLUGIN_DIR constant to load the Akismet files.