Modules: Make the generated module-headings.php pass PHPCS
The file was edited to pass phpcs in #17762, so now every `yarn build`
overwrites it with a changed version. Attempting to commit that changed
version will wind up whining about all the phpcs failures in the
generated version, and fixing them (mostly with phpcbf, some manually)
will wind up as a no-op.

This patch updates tools/build-module-headings-translations.php to fix
some of the errors and emit `phpcs:disable` for two others that isn't
worth trying to generate the way the WordPress coding standards want.