Connection: refreshing connected plugin storage on multisite networks (#17908)
While working in a multisite network, calling `switch_to_blog()` will not refresh the list of plugins, so it will still contain the data from the original blog.
Furthermore, the new blog's data will get overwritten with the data from the original site.

This PR fixes the issue by checking if the blog ID has changed every time somebody tries to pull the list of connected plugins from the `Plugin_Storage`. If the blog ID did change, the list of plugins will be pulled from the options, thus refreshing the data stored in the `Plugin_Storage`.