Packages: Update for PHP 8 testing
Packages are tested independently of WordPress, so there's no need to
try to hack a version of PHPUnit 7.5 to run under PHP 8 like WP does and
we do for the main Jetpack.

But we do still have to support PHPUnit versions from 5.7 (for PHP 5.6)
up to 9.3 (for PHP 8), so let's pull it in via yoast/phpunit-polyfills.

Also php-mock/php-mock appears to not support PHP 8. We seem to be
moving to brain/monkey for PHP function mocking anyway, so let's switch
everything over to that too.