CI: Hack around Github bug when skipping matrixed jobs
Apparently Github checks the `if` before evaluating the matrix, and if
the job is to be skipped it skips a single job with the `matrix`
contexts unexpanded instead of skipping the expanded jobs. This makes
things fail if those matrixed jobs are marked as required.

For now, let's hack around the problem by adding a step to the "detect
changed files" job to manually apply the necessary statuses when the job
is going to be skipped.