Avoid setUp/tearDown signature changes with annotations
PHPUnit 8 changed the signature for setup and teardown methods by adding
a return type declaration. And PHP 5.6 doesn't support return type
declarations, so we can't just add them.

The easiest fix is to just use `@before` and `@after` annotations to
specify the setup and teardown methods instead of overriding PHPUnit's
methods in the first place.