CI: Fix phpcs-changed
While phpcs-changed sort-of works on a shallow clone with only the two
endpoint revs fetched, it doesn't work correctly.

What it needs is to work correctly is every commit between the two
endpoints, i.e. such that `git merge-base` works. So let's fetch commits
in batches of 10 until it does.