Subscription Form: change the  default placeholder and button text values
Currently, the default email field placeholder and submit button text values shown
in the post editor do not match the values that are shown on the post after it's
published. Let's fix this inconsistency by matching the subscribePlaceholder and
submitButtonText default values to the default values used on the backend.

Fixes #18072