Subscriptions: fix comment subscription toggle behavior
There are two comment subscription settings: one to enable subscriptions
to new posts, and one to enable subscriptions to new comments. These settings
are maintained using options: stb_enabled and stc_enabled.

When these options do not exist, the settings default to 'on'. However,
because of the way that non-existent options are handled, the status of the
toggle switches for these settings does not currently match the behavior.

Two changes are necessary to fix this:

 * When a comment subscription setting is disabled, convert the false setting
   value to 0. This change will allow the option value to be updated when the
   option doesn't exist yet.

 * Remove the module name argument from the calls to getOptionValue() in
   SubscriptionsComponent. If the subscriptions module hasn't been enabled yet,
   the values of the stb_enabled and stc_enabled options are undefined, and the
   toggles display as 'off'. The default status of these settings is
   'on', and the toggles should display the actual status of the settings. Removing
   the module name argument from the getOptionValue() calls allows the status
   of the settings to be displayed correctly, regardless of whether the subscriptions
   module is enabled.