CI: Better handle failure to determine $test_root
We have a bunch of logic to try to find a copy of WordPress or
wordpress-develop. But if that fails, we wind up producing fairly
unhelpful output.

Instead, if auto-detection fails, we should output a useful message to
the user to tell them that and suggest how they might resolve the
matter.