Reorg: Enable symlinks within Docker (#18140)
The major problem with our Docker setup previously post-reorg was that package symlinks inside and outside docker didn't match up.

The main thing this does is fix that so it all just works.

package.json between repo root and the jetpack plugin root... which docker commands should live where? Added a card for that.