CI: Fix JS lint on push (#18166)
We're running the "JS Tests" workflow on both pull request and on push
to master. #18126 changed the lint job to only work right on PR, so only
run that on PR.

I note we don't run any of the other linting jobs on push to master, so
that seems reasonable.