Codesniffer: Unpin composer deps (#18209)
For libraries, it's better to use ranges so as to not send your users to
dependency hell.

Also, reconfigure renovate to not pin them again. It seems that by
default it looks at the 'type' in composer.json to decide whether it
should pin or maintain ranges, and it doesn't recognize
"phpcodesniffer-standard" as a kind it should do ranges for.