yarn: Remove broken mkdirp/minimist override (#18218)
I noticed that `yarn install` seemed to really like making this
nonsensical change to yarn.lock:
```
-minimist@0.0.8, minimist@^0.2.1:
+minimist@0.0.8, minimist@^0.2.1, minimist@^1.2.5:
   version "0.2.1"

-minimist@^1.1.1, minimist@^1.1.3, minimist@^1.2.0, minimist@^1.2.5:
+minimist@^1.1.1, minimist@^1.1.3, minimist@^1.2.0:
   version "1.2.5"
```

This turns out to be due to an override of "**/mkdirp/minimist" to
^0.2.1 in our package.json, which was happily overriding all the cases
where newer versions of mkdirp depend on minimist ^1.2.5.

It turns out that everything we have that's depending on mkdirp already
has a dependency that is compatible with said newer versions of mkdirp,
but yarn.lock had some of them still locked to an older mkdirp version.
Sigh.

So I ran `npx yarn-deduplicate --packages mkdirp` to clean that up, and
removed the broken override, and yarn.lock seems happy now. I see no
differences between mkdirp 0.5.1 and 0.5.5 that seem likely to break
anything.