Reorg: Add a dummy "jetpack.php" to the monorepo, and update docs (#18207)
People used to the old way of doing things may try to check out the
monorepo directly in their WordPress `wp-content/plugins` folder, or
they may already be doing so. That will no longer work to install
Jetpack.

To make the failure case a little more user-friendly, we add a dummy
`jetpack.php` to the monorepo root so WordPress will find a plugin. All
this plugin does is display an admin notice telling the user they're
doing it wrong and pointing them to relevant documentation on how to do
it right.

Said documentation needs to exist. So this PR also moves the existing
Jetpack readme.md to the Jetpack plugin, and creates a new readme.md for
the monorepo in the monorepo root. Once we get to the point of creating
the built "Automattic/jetpack-production" mirror repo, the
Jetpack-specific readme.md will show up there too.

Since both the Jetpack-specific readme.md and the Monorepo readme.md
want to refer to a LICENSE.txt, we also need to copy that into the
Jetpack plugin. Again, this will ensure the "Automattic/jetpack-production"
mirror repo has LICENSE.txt too once we create it.

And since there's now a fairly significant duplication of content
between the Monorepo and Jetpack in readme.md and LICENSE.txt, let's add
a pre-commit and CI check to ensure those remain in sync in the future.

And while I was looking at it, I updated a few doc refs in other files
too.