ESLint: Remove ignore for extensions/**/test/ (#18234)
As with #18233, add .eslintrc.js to the test directories to properly
configure them to handle jest's stuff.

Also clean up the existing lint errors and warnings in the test files,
which were largely whitespace and missing semicolons. A few unused
variables, most notably it looks like a test in
extensions/blocks/publicize/store/test/reducer.js wasn't actually
testing what it should have been thanks to not using the variable it had
set up to test.

And remove an obsolete temporary ignore for /_inc/search, for which the
move was apparently done in #13138.