Crowdsignal: update shortcode survey URL in using block editor (#18229)
Summary:
When embedding a Crowdsignal survey in the block editor it doesn't use the shortcode to fetch the embed code, but when the post is saved the shortcode is run. The shortcode output seems to be used for sending emails to subscribers of blogs but the survey link was incorrect because it was expecting a URL in a different, older format we don't use any more.
Ref: p1609789666043200-slack-C01CSBEN0QZ

Differential Revision: D54936-code

This commit syncs r219027-wpcom.