CI: Provide better feedback for excludelist check failures
The failure case in question is probably due to errors in maintaining
proper sorting in the file, particularly if a PR first removes an entry
then adds it back. Tell them that in the error.

I'd like to keep CI enforcing a consistent sort order on the file, as
that will help to keep the file clean for future PRs.