Stats: update dashboard widget to respect new design
Fixes #18287

WP Core recently updated the markup of the header of all dashboard widgets,
to add tools that allow one to move widgets around with the keyboard.

That change caused our implementation of the widget to break.

Our implementation is fairly fragile, as we're moving around divs with JavaScript once the page has loaded.
This commit:

- moves the widget settings to its new location.
- updates the widget css to respect the markup changes.
- moves inline CSS to the existing Dashboard widget stylesheet.
- consolidates inline JS into a single place.

Related links:
- Core change: https://core.trac.wordpress.org/changeset/48373