Pre-commit: Don't run eslint on ignored files
ESLint likes to whine if you pass it a file that's in .eslintignore, and
doesn't provide any way to override this behavior (unlike every other
linting rule).

Fortunately it uses a library called 'ignore' to process .eslintignore,
and our pre-commit check uses JS too so we can just use that library to
filter out ignored files.