CI: Fix test broken by a PHP bugfix
PHP fixed their bug 77423[1] recently, which changed the behavior of
`parse_url()` when passed a URL with an invalid userinfo part. One of
our tests was relying on the old behavior.

The simplest fix seems to be to expect either the old or new behavior
here, as the URL in question is invalid anyway.

[1]: https://bugs.php.net/bug.php?id=77423