Audio Player: Fix the firing of multiple conflicting time changes
It was seen that if you click around on the time line of the audio
player, that from time to time, the click will be ignored and the audio
will jump back to the previous position.

This change updates the the options passed to `throttle` as the problem
seemed to be caused by the prop changing every second and a click
happening within the second, this led to two calls to the throttled
function, on the leading and trailing edge of the time period.

It's probably not strictly necessary, but it also cancels the throttled
function in the unmount callback of the `useEffect`.