Fix "Build artifacts" job
When we removed .svnignore, that caused the job to start trying to
include all of node_modules and such in the artifact which is completely
not what we want.

We already have the logic to produce exactly what we do want for the
mirror repos, so refactor the existing script into "build" and "push"
parts so "Build artifacts" can use just the first part.