Add a repo for the built projects to plugins' composer.json
The monorepo mirroring works fine for non-production builds, because it
symlinks the whole source directory.

But production builds run composer with --prefer-dist, which does
`git archive` or something similar to get the distribution version of
the package. And that doesn't handle our custom production-include
attribute. Since we already have these directories all prepared for the
production build, we can just use them instead.