Debug-Helper: Add composer.json so it will be mirrored (#18379)
A miscommunication on #18317 led to the file not being included, but
then thinking it had been #18355 removed the ability to mirror without a
composer.json.

We're all agreed we want a composer.json here, so let's add it now. We
can always edit it again after p9dueE-2l9-p2 reaches a decision.