Fix search inputs to work independently
Previously the 'X' close link for one search input was closing the
results for both search inputs. This change makes the display of the
'X' close link and the clearing/closing of the search inputs work
independently for both inputs.

The change also correctly clears out the input values when clicking on
the close links which was previously throwing a JS error.